apps/requirements.txt
dash
flask
haystack-ai
huggingface_hub